src/backend/reg_cop_btc/pyproject.toml
[project]
name = "reg-cop-btc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.7",
    "python-bitcoinrpc>=1.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]
